convert Mission components from Javascript to Typescript
